Refactor: Organize model classes into separate modules under models package
