+++ models/__init__.py
@@ -1,0 +1,10 @@
+from flask_sqlalchemy import SQLAlchemy
+
+db = SQLAlchemy()
+
+from .event import Event
+from .family import Family
+from .link import Link
+from .member import Member
+from .relationship import Relationship
+from .user import User
